@@ -6,23 +6,23 @@
 		"start": "ng serve -o",
 		"build": "ng build",
 		"build:href": "ng build --base-href ./",
-		"build:github": "npm run delete:docs && npm run build:href && npm run copy:dist && npm run copy:docs",
+		"build:github": "npm run delete:docs && npm run build:href && npm run copy:dist && npm run copy:dist",
 		"watch": "ng build --watch --configuration development",
 		"test": "ng test",
 		"delete:docs": "del docs",
-		"copy:dist": "copyfiles -u 2 dist/my-dashboard/*.txt ./docs",
+		"copy:dist": "copyfiles -u 2 dist/my-dashboard/**/* ./docs",
 		"copy:docs": "copyfiles dist/my-dashboard/browser/* ./docs -f"
 	},
 	"private": true,
 	"dependencies": {
-		"@angular/animations": "^17.3.0",
-		"@angular/common": "^17.3.0",
-		"@angular/compiler": "^17.3.0",
-		"@angular/core": "^17.3.0",
-		"@angular/forms": "^17.3.0",
-		"@angular/platform-browser": "^17.3.0",
-		"@angular/platform-browser-dynamic": "^17.3.0",
-		"@angular/router": "^17.3.0",
+		"@angular/animations": "^18.0.5",
+		"@angular/common": "^18.0.5",
+		"@angular/compiler": "^18.0.5",
+		"@angular/core": "^18.0.5",
+		"@angular/forms": "^18.0.5",
+		"@angular/platform-browser": "^18.0.5",
+		"@angular/platform-browser-dynamic": "^18.0.5",
+		"@angular/router": "^18.0.5",
 		"copyfiles": "^2.4.1",
 		"del-cli": "^5.1.0",
 		"rxjs": "~7.8.0",
@@ -30,9 +30,9 @@
 		"zone.js": "~0.14.3"
 	},
 	"devDependencies": {
-		"@angular-devkit/build-angular": "^17.3.2",
-		"@angular/cli": "^17.3.2",
-		"@angular/compiler-cli": "^17.3.0",
+		"@angular-devkit/build-angular": "^18.0.6",
+		"@angular/cli": "^18.0.6",
+		"@angular/compiler-cli": "^18.0.5",
 		"@types/jasmine": "~5.1.0",
 		"autoprefixer": "^10.4.19",
 		"jasmine-core": "~5.1.0",
