@@ -1,8 +1,8 @@
-import { ApplicationConfig, importProvidersFrom } from '@angular/core';
+import { ApplicationConfig } from '@angular/core';
 import { provideRouter, withViewTransitions } from '@angular/router';
 
 import { routes } from './app.routes';
-import { HttpClientModule } from '@angular/common/http';
+import { provideHttpClient, withInterceptorsFromDi } from '@angular/common/http';
 
 export const appConfig: ApplicationConfig = {
 	providers: [
@@ -15,6 +15,7 @@ export const appConfig: ApplicationConfig = {
 				}*/
 			})
 		),
-		importProvidersFrom (HttpClientModule),
+		// importProvidersFrom (HttpClientModule),
+		provideHttpClient (withInterceptorsFromDi ()),
 	]
 };
