@@ -4,7 +4,7 @@ import { RouterOutlet } from '@angular/router';
 @Component({
 	selector: 'app-root',
 	standalone: true,
-	imports: [RouterOutlet],
+	imports: [ RouterOutlet ],
 	templateUrl: './app.component.html',
 	styleUrl: './app.component.css'
 })
