@@ -1,8 +1,8 @@
 import { Component } from '@angular/core';
 import { RouterModule } from '@angular/router';
 
-// import { SideMenuComponent } from '@shared/side-menu/side-menu.component';
-import { SideMenuComponent } from '../shared/side-menu/side-menu.component';
+import { SideMenuComponent } from '@shared/side-menu/side-menu.component';
+// import { SideMenuComponent } from '../shared/side-menu/side-menu.component';
 
 @Component({
 	selector: 'dashboard',
