@@ -1,15 +1,15 @@
 import { Component, WritableSignal, signal } from '@angular/core';
 import { CommonModule } from '@angular/common';
 
-// import { Grade } from '@interfaces/grade.type';
-import { Grade } from '../../../interfaces/grade.type';
+import { Grade } from '@interfaces/grade.type';
+// import { Grade } from '../../../interfaces/grade.type';
 
-// import { TitleComponent } from '@shared/title/title.component';
-import { TitleComponent } from '../../../shared/title/title.component';
+import { TitleComponent } from '@shared/title/title.component';
+// import { TitleComponent } from '../../../shared/title/title.component';
 
 @Component({
 	standalone: true,
-	imports: [CommonModule, TitleComponent],
+	imports: [ CommonModule, TitleComponent ],
 	templateUrl: './control-flow.component.html',
 	styles: ``
 })
