@@ -1,10 +1,10 @@
 import { CommonModule } from '@angular/common';
 import { Component } from '@angular/core';
 
-// import { HeavyLoadersFastComponent } from '@shared/heavy-loaders/heavy-loaders-fast.component';
-// import { TitleComponent } from '@shared/title/title.component';
-import { HeavyLoadersFastComponent } from '../../../shared/heavy-loaders/heavy-loaders-fast.component';
-import { TitleComponent } from '../../../shared/title/title.component';
+import { HeavyLoadersFastComponent } from '@shared/heavy-loaders/heavy-loaders-fast.component';
+import { TitleComponent } from '@shared/title/title.component';
+// import { HeavyLoadersFastComponent } from '../../../shared/heavy-loaders/heavy-loaders-fast.component';
+// import { TitleComponent } from '../../../shared/title/title.component';
 
 @Component({
 	standalone: true,
