@@ -1,14 +1,14 @@
 import { CommonModule } from '@angular/common';
 import { Component } from '@angular/core';
 
-// import { HeavyLoadersSlowComponent } from '@shared/heavy-loaders/heavy-loaders-slow.component';
-// import { TitleComponent } from '@shared/title/title.component';
-import { HeavyLoadersSlowComponent } from '../../../shared/heavy-loaders/heavy-loaders-slow.component';
-import { TitleComponent } from '../../../shared/title/title.component';
+import { HeavyLoadersSlowComponent } from '@shared/heavy-loaders/heavy-loaders-slow.component';
+import { TitleComponent } from '@shared/title/title.component';
+// import { HeavyLoadersSlowComponent } from '../../../shared/heavy-loaders/heavy-loaders-slow.component';
+// import { TitleComponent } from '../../../shared/title/title.component';
 
 @Component({
 	standalone: true,
-	imports: [CommonModule, HeavyLoadersSlowComponent, TitleComponent],
+	imports: [ CommonModule, HeavyLoadersSlowComponent, TitleComponent ],
 	templateUrl: './defer-views.component.html',
 	styles: ``
 })
