@@ -2,11 +2,11 @@ import { CommonModule } from '@angular/common';
 import { Component, inject } from '@angular/core';
 import { RouterModule } from '@angular/router';
 
-// import { UsersService } from '@services/users.service';
-import { UsersService } from '../../../services/users.service';
+import { UsersService } from '@services/users.service';
+// import { UsersService } from '../../../services/users.service';
 
-// import { TitleComponent } from '@shared/title/title.component';
-import { TitleComponent } from '../../../shared/title/title.component';
+import { TitleComponent } from '@shared/title/title.component';
+// import { TitleComponent } from '../../../shared/title/title.component';
 
 @Component({
 	standalone: true,
