@@ -1,12 +1,12 @@
 import { CommonModule } from '@angular/common';
 import { Component } from '@angular/core';
 
-// import { TitleComponent } from '@shared/title/title.component';
-import { TitleComponent } from '../../../shared/title/title.component';
+import { TitleComponent } from '@shared/title/title.component';
+// import { TitleComponent } from '../../../shared/title/title.component';
 
 @Component({
 	standalone: true,
-	imports: [ CommonModule, TitleComponent],
+	imports: [ CommonModule, TitleComponent ],
 	template: `
 		<shared-title title="View Transition 2" />
 
