@@ -1,5 +1,5 @@
-// import { User } from '@interfaces/req-response.interface';
-import { User } from './req-response.interface';
+import { User } from '@interfaces/req-response.interface';
+// import { User } from './req-response.interface';
 
 export interface State {
 	users: User[];
